backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
langchain-openai
langchain-core
snowleopard
